refactor(api): extract Conversation/Message TypedDict definitions into model_dict_types module
